some file restructuring, the tx and rx are now inside the peermanager itself
